Database Object Explorer

DBOE facilitates the navigation of SQL-based database engines by means of metadata. Engines tested to date include Microsoft SQL Server and MySQL.

Active bindings

connection.list

Sets or returns a list of saved connections. When providing a list, it should be names by database.

Methods

Public methods

DBOE$new()

DBOE$get.metadata()

DBOE$make.virtual_database()

DBOE$clone()

Method new()

Initialize the class object

Usage

DBOE$new()

Returns

The class object, invisibly

Method get.metadata()

$get.metadata() retrieves metadata information for the database pointed to by argument conns. Once metadata has been retrieved, metadata can be accessed for tables, views, and stored procedures using the following access method: <DBOE obj>$<database name>$<table/view/proc name>

Usage

DBOE$get.metadata(..., include.procs = FALSE, chatty = FALSE)

Arguments

...

dots_list: one or more DBI/ODBC connection objects.

include.procs

(logical | FALSE) Should stored procedures be included in the metatdata retrieval?

chatty

(logical | FALSE) Should additional execution messages be sent to the console?

Returns

The class object, invisibly

Method make.virtual_database()

$make.virtual_database creates a set of tbl objects in an environment.

Usage

DBOE$make.virtual_database(
  conn,
  ...,
  target_env = rlang::caller_env(),
  sch = "dbo"
)

Arguments

conn

The name of a metadata environment (created after calling $get.metadata() )

...

Names or patterns of objects to retrieve to link from source.

target_env

The environment object where created objects should be stored

sch

(Optional) The target schema (e.g., dbo) on which to filter internal search results

Returns

An assignable environment object with DBI-sourced tbls

Method clone()

The objects of this class are cloneable with this method.

Usage

DBOE$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.